# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/agents/{agentID}/runs:
    get:
      operationId: listAgentRuns
      description: List durable Agent runs visible to the current Console account. Queued runs remain pending until a trusted execution worker is available.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AgentID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Agent run history, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunsPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { $ref: '#/components/responses/InternalError' } }
    post:
      operationId: createAgentRun
      description: Accept an Agent prompt into the durable run queue. The request does not call a provider or execute repository code inline.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AgentID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateAgentRunRequest' } } } }
      responses: { '202': { description: Run accepted, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/ValidationError' }, '500': { $ref: '#/components/responses/InternalError' } }
  /v1/agents/{agentID}/runs/{runID}:
    get:
      operationId: getAgentRun
      description: Get one durable Agent run and any worker-produced steps, output, or changes.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AgentID' }, { $ref: '#/components/parameters/RunID' }]
      responses: { '200': { description: Agent run, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { $ref: '#/components/responses/InternalError' } }
  /v1/agents/{agentID}/runs/{runID}/cancel:
    post:
      operationId: cancelAgentRun
      description: Cancel a queued or running Agent run. Only project owners and admins may cancel work.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AgentID' }, { $ref: '#/components/parameters/RunID' }]
      responses: { '200': { description: Cancelled Agent run, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '500': { $ref: '#/components/responses/InternalError' } }
  /v1/agents/{agentID}/runs/{runID}/logs:
    get:
      operationId: listAgentRunLogs
      description: Read bounded execution logs emitted by the trusted Agent worker. Use after with the returned sequence number for incremental polling.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/AgentID' }, { $ref: '#/components/parameters/RunID' }, { $ref: '#/components/parameters/Limit' }, { name: after, in: query, schema: { type: integer, format: int64, minimum: 0, default: 0 } }]
      responses: { '200': { description: Agent run logs, content: { application/json: { schema: { $ref: '#/components/schemas/AgentRunLogsPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { $ref: '#/components/responses/InternalError' } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes are users.read/write, databases.read/write, storage.read/write, functions.read/write, sites.read/write, webhooks.read/write, and realtime.read; never send this header through the Console browser bridge. }
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, schema: { type: string, format: uuid } }
...
    AgentID: { name: agentID, in: path, required: true, schema: { type: string, format: uuid } }
    RunID: { name: runID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Invalid JSON or parameter, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    PayloadTooLarge: { description: Request body exceeds the endpoint's configured limit (JSON requests are bounded separately from multipart uploads), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json except for file upload multipart/form-data endpoints, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    Pagination: { type: object, required: [limit, next_cursor], properties: { limit: { type: integer }, next_cursor: { type: [string, 'null'], format: uuid } } }
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    AgentRunStep: { type: object, additionalProperties: false, required: [id, type, label, target, status], properties: { id: { type: string, maxLength: 128 }, type: { type: string, enum: [read, edit, search, command, check] }, label: { type: string, minLength: 1, maxLength: 256 }, target: { type: string, minLength: 1, maxLength: 2000 }, status: { type: string, enum: [pending, done] } } }
    AgentRunChange: { type: object, additionalProperties: false, required: [path, additions, deletions, status], properties: { path: { type: string, minLength: 1, maxLength: 1000 }, additions: { type: integer, minimum: 0 }, deletions: { type: integer, minimum: 0 }, status: { type: string, enum: [added, modified] } } }
    AgentRun: { type: object, additionalProperties: false, required: [id, agent_id, project_id, prompt, status, steps, changes, queued_at, created_at, updated_at], properties: { id: { type: string, format: uuid }, agent_id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, prompt: { type: string, minLength: 1, maxLength: 20000 }, status: { type: string, enum: [queued, running, completed, failed, cancelled] }, output_text: { type: [string, 'null'], maxLength: 100000 }, error_message: { type: [string, 'null'], maxLength: 4000 }, steps: { type: array, maxItems: 256, items: { $ref: '#/components/schemas/AgentRunStep' } }, changes: { type: array, maxItems: 256, items: { $ref: '#/components/schemas/AgentRunChange' } }, created_by_account_id: { type: [string, 'null'], format: uuid }, queued_at: { type: string, format: date-time }, started_at: { type: [string, 'null'], format: date-time }, finished_at: { type: [string, 'null'], format: date-time }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
    AgentRunLog: { type: object, additionalProperties: false, required: [id, run_id, project_id, sequence, level, message, created_at], properties: { id: { type: string, format: uuid }, run_id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, sequence: { type: integer, format: int64, minimum: 1 }, level: { type: string, enum: [debug, info, warn, error] }, message: { type: string, minLength: 1, maxLength: 16000 }, created_at: { type: string, format: date-time } } }
    CreateAgentRunRequest: { type: object, additionalProperties: false, required: [prompt], properties: { prompt: { type: string, minLength: 1, maxLength: 20000, description: Prompt queued for a trusted execution worker. } } }
    AgentRunResponse: { type: object, additionalProperties: false, required: [run], properties: { run: { $ref: '#/components/schemas/AgentRun' } } }
    AgentRunsPage: { type: object, additionalProperties: false, required: [runs, pagination], properties: { runs: { type: array, items: { $ref: '#/components/schemas/AgentRun' } }, pagination: { $ref: '#/components/schemas/Pagination' } } }
    AgentRunLogsPage: { type: object, additionalProperties: false, required: [logs, pagination], properties: { logs: { type: array, items: { $ref: '#/components/schemas/AgentRunLog' } }, pagination: { $ref: '#/components/schemas/Pagination' } } }
...